% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Treggae_signatures.R
\name{get_signature}
\alias{get_signature}
\title{Get a single signatures}
\usage{
get_signature(signature_name)
}
\arguments{
\item{signature_name}{Character string, corresponding to the name of the
signature one wants to access}
}
\value{
A character vector
}
\description{
Gets a single signature among the ones provided in Treggae
}
\examples{

get_signature("Skin_Tregs_signature")

}
