% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Treggae_signatures.R
\name{get_signatures}
\alias{get_signatures}
\title{Get all signatures}
\usage{
get_signatures()
}
\value{
A list object
}
\description{
Obtain a list with all signatures contained in Treggae
}
\examples{
all_sigs <- get_signatures()
all_sigs$Skin_Tregs_signature

}
