% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Treggae_signatures.R
\name{list_signatures}
\alias{list_signatures}
\title{List all signatures}
\usage{
list_signatures()
}
\value{
Invisible NULL. Prints a summary as a message
}
\description{
List all signatures available in Treggae with a compact summary
}
\examples{
list_signatures()

}
