% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Treggae-pkg.R
\docType{package}
\name{Treggae-pkg}
\alias{Treggae}
\alias{Treggae-package}
\alias{Treggae-pkg}
\title{Treggae}
\description{
\code{Treggae} is a package that ...
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
