% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Treggae-pkg.R
\docType{package}
\name{Treggae-pkg}
\alias{Treggae}
\alias{Treggae-package}
\alias{Treggae-pkg}
\title{Treggae}
\description{
\code{Treggae} is a package that ...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/imbeimainz/Treggae/}
  \item Report bugs at \url{https://github.com/imbeimainz/Treggae/issues}
}

}
\author{
\strong{Maintainer}: Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})

}
\keyword{internal}
